Update CreateJobFilter.cs
BackgroundJob is always null when using batch functionality from Hangfire.Pro packages. For this reason we cannot use TagAttribute to decorate the job method.
This change allows to not fail the job in such case when there is no TagAttribute decorating the method and backward compatibility with users who are using it.